проверка потолков

+++ rotate.py
@@ -7,9 +7,9 @@ y'=x\sin\theta+y\cos\theta.
 '''
 
 def hRotatePoint(pts, num, angle):
-    angle = radians(float(int(angle)*int(num)))
+    angle = radians(float(angle)*float(num))
     rezPoint = []
-    rezPoint.append(float(pts[0])*cos(angle) - float(pts[1])*sin(angle))
-    rezPoint.append(float(pts[1])*cos(angle) + float(pts[0])*sin(angle))
-    rezPoint.append(float(pts[2]))
+    rezPoint.append(pts[0]*cos(angle) - pts[1]*sin(angle))
+    rezPoint.append(pts[1]*cos(angle) + pts[0]*sin(angle))
+    rezPoint.append(pts[2])
     return rezPoint
